# Deep-Research/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateProfileInput {
  email?: string;
  /** @minLength 8 */
  password?: string;
}

export interface ResearchInput {
  /** @minLength 10 */
  query: string;
}

export type StartedResearchStatus = typeof StartedResearchStatus[keyof typeof StartedResearchStatus];


export const StartedResearchStatus = {
  planning: 'planning',
  searching: 'searching',
  reading: 'reading',
  synthesizing: 'synthesizing',
  verifying: 'verifying',
  formatting: 'formatting',
  done: 'done',
  paused: 'paused',
  failed: 'failed',
} as const;

export interface StartedResearch {
  research_id: string;
  id: string;
  query: string;
  status: StartedResearchStatus;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  queued: 'queued',
  active: 'active',
  done: 'done',
} as const;

export interface Agent {
  name: string;
  detail: string;
  count: string;
  status: AgentStatus;
  progress: number;
}

export interface Source {
  type: string;
  title: string;
  source: string;
  progress: number;
}

export interface ResearchSource {
  researchId: string;
  researchQuery: string;
  status: string;
  type: string;
  title: string;
  source: string;
  progress: number;
}

export type PaginatedSourcesPagination = {
  total: number;
  page: number;
  limit: number;
  totalPages: number;
};

export interface PaginatedSources {
  data: ResearchSource[];
  pagination: PaginatedSourcesPagination;
}

export interface RecentResearch {
  id: string;
  title: string;
  meta: string;
  status: string;
}

export type ResearchStatus = typeof ResearchStatus[keyof typeof ResearchStatus];


export const ResearchStatus = {
  planning: 'planning',
  searching: 'searching',
  reading: 'reading',
  synthesizing: 'synthesizing',
  verifying: 'verifying',
  formatting: 'formatting',
  done: 'done',
  paused: 'paused',
  failed: 'failed',
} as const;

export interface Research {
  id: string;
  query: string;
  status: ResearchStatus;
  progress: number;
  elapsedMinutes: number;
  verificationScore: number;
  sourcesCount: number;
  claimsChecked: number;
  agents: Agent[];
  sources: Source[];
  recent: RecentResearch[];
  summary: string;
  report: string;
  createdAt: string;
  updatedAt: string;
}

export interface WorkspaceSummary {
  activeResearch: number;
  completedReports: number;
  sourcesRead: number;
  contextUsed: number;
}

export interface ResearchUpdateInput {
  /** @minLength 1 */
  query: string;
}

export interface DeleteResearchResult {
  ok: boolean;
}

export interface WorkspaceUsage {
  totalTokensUsed: number;
  totalSourcesIndexed: number;
  totalReports: number;
  maxContextLimit: number;
  usedContextPct: number;
}

export interface PurgeCacheResult {
  purgedThreads: number;
  totalTokensUsed: number;
  totalSourcesIndexed: number;
  totalReports: number;
  maxContextLimit: number;
  usedContextPct: number;
}

export type ListSourcesParams = {
/**
 * Page number (1-based)
 * @minimum 1
 */
page?: number;
/**
 * Number of sources per page
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};