# Deep-Research/lib/api-zod/src/generated/types/updateProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileInput {
  email?: string;
  /** @minLength 8 */
  password?: string;
}